putting fork in project.
Being unable to find funding at this time to pay
for continued development & for infra this project
is coming to an end. Best wishes to anyone who is able
to continue this. Please dont bother sending me patches
just fork the project and run with it how ever you see fit!

Michael Rice